Fix for latest Xcode versions and Travis builds.

+++ Device Monitor/Classes/Controllers/FileSystem/Details/DMFileSystemImageViewController.m
@@ -144,7 +144,7 @@
     }
     
     image      = [ UIImage imageWithContentsOfFile: _file.path ];
-    label.text = [ NSString stringWithFormat: @"%u x %u", ( NSUInteger )image.size.width, ( NSUInteger )image.size.height ];
+    label.text = [ NSString stringWithFormat: @"%lu x %lu", ( unsigned long )image.size.width, ( unsigned long )image.size.height ];
     
     item                                    = [ [ UIBarButtonItem alloc ] initWithCustomView: label ];
     infos                                   = [ [ UIBarButtonItem alloc ] initWithBarButtonSystemItem: UIBarButtonSystemItemSearch target: self action: @selector( showInfos: ) ];
